Implement Algorand-style VRF sortition for block producer selection
SECURITY: Replaces deterministic shared-hash leader selection with per-node
VRF evaluation. Each masternode evaluates VRF(secret_key, input) locally to
determine eligibility — only the selected node knows it's the leader until
it reveals the VRF proof.

Changes:
- Block production: VRF self-selection using sampling weights (§5.2)
- Block validation: Verify VRF proof, score, and proposer eligibility
- Best-proposal selection: Lower VRF score wins when multiple proposals
- Timeout fallback: Exponential threshold relaxation (2^attempt)
- BlockCache: Add get_by_height() for proposal comparison
- ValidatorInfo.weight: usize -> u64 (match sampling_weight return type)
- select_leader() kept in masternode_registry as legacy fallback

+++ src/blockchain.rs
@@ -2969,13 +2969,13 @@ impl Blockchain {
 
                 // Get peer's tier weight (default to Bronze if not found)
                 let peer_weight = match self.masternode_registry.get(peer_ip).await {
-                    Some(info) => info.masternode.tier.sampling_weight() as u64,
+                    Some(info) => info.masternode.tier.sampling_weight(),
                     None => {
                         tracing::debug!(
                             "Peer {} not in masternode registry, using Bronze weight",
                             peer_ip
                         );
-                        crate::types::MasternodeTier::Bronze.sampling_weight() as u64
+                        crate::types::MasternodeTier::Bronze.sampling_weight()
                     }
                 };
 
@@ -6183,8 +6183,8 @@ impl Blockchain {
         let mut total_responding_weight = 0u64;
         for peer_ip in peer_tips.keys() {
             let peer_weight = match self.masternode_registry.get(peer_ip).await {
-                Some(info) => info.masternode.tier.sampling_weight() as u64,
-                None => crate::types::MasternodeTier::Bronze.sampling_weight() as u64,
+                Some(info) => info.masternode.tier.sampling_weight(),
+                None => crate::types::MasternodeTier::Bronze.sampling_weight(),
             };
             total_responding_weight += peer_weight;
             if consensus_peers.contains(peer_ip) {
